+++ src/main/java/insane96mcp/enhancedai/mixin/ShulkerBulletMixin.java
@@ -1,6 +1,6 @@
 package insane96mcp.enhancedai.mixin;
 
-import insane96mcp.enhancedai.modules.shulker.feature.ShulkerBullets;
+import insane96mcp.enhancedai.modules.shulker.ShulkerBullets;
 import net.minecraft.world.effect.MobEffectInstance;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.projectile.Projectile;