+++ src/main/java/insane96mcp/enhancedai/modules/enderman/ItemDisruption.java
@@ -1,4 +1,4 @@
-package insane96mcp.enhancedai.modules.enderman.feature;
+package insane96mcp.enhancedai.modules.enderman;
 
 import insane96mcp.enhancedai.modules.Modules;
 import insane96mcp.insanelib.base.Feature;