+++ src/main/java/insane96mcp/enhancedai/modules/pets/snowgolem/SnowGolems.java
@@ -1,8 +1,7 @@
-package insane96mcp.enhancedai.modules.pets.feature;
+package insane96mcp.enhancedai.modules.pets.snowgolem;
 
 import insane96mcp.enhancedai.EnhancedAI;
 import insane96mcp.enhancedai.modules.Modules;
-import insane96mcp.enhancedai.modules.pets.ai.EARangedSnowGolemAttackGoal;
 import insane96mcp.enhancedai.setup.NBTUtils;
 import insane96mcp.insanelib.base.Feature;
 import insane96mcp.insanelib.base.Label;